Restructure to src/ layout, add build system, update all imports and tests
