+++ conftest.py
@@ -1,0 +1,3 @@
+import sys
+import os
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src", "resume_scanner"))
