+++ pyproject.toml
@@ -22,3 +22,10 @@ dev = [
     "pytest-cov>=7.1.0",
     "pytest-mock>=3.15.1",
 ]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/resume_scanner"]
