+++ src/resume_scanner/app/routers/analyze.py
@@ -7,7 +7,7 @@ from dotenv import load_dotenv
 from fastapi import APIRouter, File, HTTPException, UploadFile
 from openai import OpenAI, OpenAIError
 
-from app.models.schemas import AnalysisResponse, AnalysisResult, ResumeTextRequest
+from resume_scanner.app.models.schemas import AnalysisResponse, AnalysisResult, ResumeTextRequest
 
 load_dotenv()
 client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
