+++ src/resume_scanner/app/routers/ats.py
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 from fastapi import APIRouter, File, Form, HTTPException, UploadFile
 from openai import OpenAI, OpenAIError
 
-from app.models.schemas import ATSRequest, ATSResponse, ATSResult
+from resume_scanner.app.models.schemas import ATSRequest, ATSResponse, ATSResult
 
 load_dotenv()
 client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
