+++ src/resume_scanner/app/routers/resume.py
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 from fastapi import APIRouter, File, HTTPException, UploadFile
 from openai import OpenAI, OpenAIError
 
-from app.models.schemas import ResumeAnalysisResponse, ResumeTextRequest, RoleMatch
+from resume_scanner.app.models.schemas import ResumeAnalysisResponse, ResumeTextRequest, RoleMatch
 
 load_dotenv()
 client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))
