+++ tests/test_health.py
@@ -1,7 +1,7 @@
 import pytest
 from fastapi.testclient import TestClient
 
-from main import app
+from resume_scanner.main import app
 
 client = TestClient(app)
